Reworking serializers to work with DataContract/DataMember
